Split shared frontend surfaces
